app/requirements.txt
# 🚀 Blacklist Flask Application Dependencies
# Core Flask Dependencies
Flask==2.3.3
Werkzeug==2.3.7
Jinja2==3.1.2
MarkupSafe==2.1.3

# Database & Connection Pooling
psycopg2-binary==2.9.7
redis==4.6.0

# Data Processing & Analysis
pandas>=2.2.0  # Python 3.13 compatible
numpy>=1.26.0  # Python 3.13 compatible
openpyxl==3.1.2

# Web Scraping & HTTP Requests
requests==2.31.0
selenium==4.15.2
beautifulsoup4==4.12.2

# Security & Authentication
cryptography==44.0.0
bcrypt==4.0.1
Flask-Login==0.6.3          # User session management
Flask-WTF==1.2.1            # CSRF protection
Flask-Limiter==3.5.0        # Rate limiting / DOS protection

# API & JSON Processing
marshmallow==3.20.1
jsonschema==4.19.1

# Utilities
python-dotenv==1.0.0
click==8.1.7
pytz==2023.3
schedule==1.2.0

# Production Server
gunicorn==21.2.0
gevent==25.9.1

# System Monitoring & Metrics
psutil==5.9.5                    # CRITICAL: Required for metrics routes
prometheus-client==0.18.0        # Metrics collection

# Structured Logging (Phase 1.2)
structlog==23.2.0                # Structured logging for decision tracking

# Development & Testing
pytest==7.4.3
pytest-cov==4.1.0
pytest-mock==3.12.0
pytest-flask==1.3.0
coverage==7.3.2
black==23.9.1
flake8==6.1.0
